requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
supabase==2.3.0
sentence-transformers==3.3.1
PyPDF2==3.0.1
numpy==1.24.3